Refactor project structure (added iride_staramr_results package, combined irida_api and staramr_writer funtionality).
